refactor: ESLintのルールに準拠
