+++ src/components/ComponentPreview/ComponentPreview.tsx
@@ -1,7 +1,9 @@
 import React, { type CSSProperties } from 'react';
 import { Cluster } from 'smarthr-ui';
-import type { Gap, SeparateGap } from 'smarthr-ui/lib/types';
+
 import WrapperBase from './WrapperBase';
+
+import type { Gap, SeparateGap } from 'smarthr-ui/lib/types';
 
 export type ComponentPreviewProps = {
   children: React.ReactNode;
@@ -16,6 +18,7 @@ export default function ComponentPreview({
 }: ComponentPreviewProps) {
   return (
     <WrapperBase style={{ padding: '1.5em' }}>
+      {/* eslint-disable-next-line smarthr/best-practice-for-layouts */}
       <Cluster gap={gap} align={align}>
         {children}
       </Cluster>
