+++ src/components/article/ComponentCaptures.astro
@@ -17,7 +17,6 @@ const storyGroups = await fetchComponentCaptures();
                 <a href={`/products/components/${storyKind.componentPath}/`}>
                   <div>
                     {/* この画像はリンクテキストと同等の内容なので、altは空が適切 */}
-                    {/* eslint-disable-next-line smarthr/a11y-image-has-alt-attribute */}
                     <img width={300} height={200} src={`/thumbnails/component-stories/${storyKind.thumbnailFileName}`} alt="" />
                   </div>
                   <p>{storyKind.displayName}</p>
