+++ src/content/articles/operational-guideline/page-template/style-template.mdx
@@ -6,7 +6,7 @@ order: 1
 
 import { Image } from 'astro:assets'
 import { Button, Cluster, FaPlusCircleIcon, Table, Td, Th, Text } from 'smarthr-ui'
-import ComponentPreview from '@/components/ComponentPreview'
+import ComponentPreview from '@/components/ComponentPreview/ComponentPreview'
 import DoAndDont from '@/components/article/DoAndDont.astro'
 import TableWrapper from '@/components/article/TableWrapper.astro'
