+++ src/content/articles/products/components/dialog/index.mdx
@@ -5,7 +5,7 @@ description: 'ダイアログボックス（もしくは単に「ダイアログ
 
 import ComponentPropsTable from '@/components/article/ComponentPropsTable.astro'
 import ComponentStory from '@/components/article/ComponentStory.astro'
-import ComponentPreview from '@/components/ComponentPreview'
+import ComponentPreview from '@/components/ComponentPreview/ComponentPreview'
 import {
   ActionDialog,
   Button,
