+++ src/content/articles/products/components/flash-message/index.mdx
@@ -5,7 +5,7 @@ description: 'FlashMessageは非推奨です。ユーザーに操作の結果を
 
 import ComponentPropsTable from '@/components/article/ComponentPropsTable.astro'
 import ComponentStory from '@/components/article/ComponentStory.astro'
-import ComponentPreview from '@/components/ComponentPreview'
+import ComponentPreview from '@/components/ComponentPreview/ComponentPreview'
 import Warning from './_components/Warning'
 
 <Warning />
