+++ src/content/articles/products/components/icon.mdx
@@ -5,7 +5,7 @@ description: 'アイデアやアクションを表すための視覚的な要素
 
 import ComponentPropsTable from '@/components/article/ComponentPropsTable.astro'
 import ComponentStory from '@/components/article/ComponentStory.astro'
-import ComponentPreview from '@/components/ComponentPreview'
+import ComponentPreview from '@/components/ComponentPreview/ComponentPreview'
 import {
   FaAddressBookIcon,
   FaAddressCardIcon,
