+++ src/content/articles/products/components/input/_components/SampleInputWithPrefixIcon.tsx
@@ -1,5 +1,6 @@
 import { FaLockIcon, Input } from 'smarthr-ui';
 
 export default function SampleInputWithPrefixIcon() {
+  // eslint-disable-next-line smarthr/a11y-input-in-form-control
   return <Input name="exampleprefixicon" type="password" defaultValue="password" prefix={<FaLockIcon />} />;
 }
