+++ src/content/articles/products/components/notification-bar/index.mdx
@@ -2,7 +2,7 @@
 title: 'NotificationBar'
 description: '操作の結果などをシステムからの通知として表示するためのコンポーネントです。'
 ---
-import ComponentPreview from '@/components/ComponentPreview'
+import ComponentPreview from '@/components/ComponentPreview/ComponentPreview'
 import ComponentStory from '@/components/article/ComponentStory.astro'
 import ComponentPropsTable from '@/components/article/ComponentPropsTable.astro'
 import VideoEmbed from '@/components/article/VideoEmbed.astro'
