+++ src/content/articles/products/components/tooltip/index.mdx
@@ -16,7 +16,7 @@ import {
   FaPlusCircleIcon,
 } from 'smarthr-ui'
 import { Button, Cluster, Tooltip, Text } from 'smarthr-ui'
-import ComponentPreview from '@/components/ComponentPreview'
+import ComponentPreview from '@/components/ComponentPreview/ComponentPreview'
 import ComponentStory from '@/components/article/ComponentStory.astro'
 import ComponentPropsTable from '@/components/article/ComponentPropsTable.astro'
