+++ src/content/articles/products/design-patterns/delete-dialog/index.mdx
@@ -4,7 +4,7 @@ description: '削除操作をするときのアクションダイアログを定
 order: 9
 ---
 
-import ComponentPreview from '@/components/ComponentPreview'
+import ComponentPreview from '@/components/ComponentPreview/ComponentPreview'
 
 削除操作をするときのアクションダイアログを定義します。
