+++ src/content/articles/products/design-tokens/shadow/index.mdx
@@ -4,7 +4,7 @@ description: '影のトークンは、要素の周りに落とす影（シャド
 order: 7
 ---
 
-import ComponentPreview from '@/components/ComponentPreview'
+import ComponentPreview from '@/components/ComponentPreview/ComponentPreview'
 import ShadowSample from './_components/ShadowSample'
 
 影のトークンは、要素の周りに落とす影（シャドウ）の大きさを定義します。
