+++ src/remark/remark-code-block.ts
@@ -1,13 +1,13 @@
+import { type Visitor, visit } from 'unist-util-visit';
+
 import type { Code, Paragraph } from 'mdast';
 import type { Node } from 'unist';
-import { type Visitor, visit } from 'unist-util-visit';
 
 /**
  * 「```tsx editable」のようなコードブロックから、コードと言語・メタ情報を取り出すプラグイン
  * 参考: https://blog.mono0x.net/2023/07/10/astro-syntax-highlight-with-title/
  */
-const remarkCodeBlock = () => {
-  return (tree: Node) => {
+const remarkCodeBlock = () => (tree: Node) => {
     const visitor: Visitor<Code> = (node, index, parent) => {
       if (!parent || index === undefined) {
         return;
@@ -31,6 +31,5 @@ const remarkCodeBlock = () => {
 
     visit(tree, 'code', visitor);
   };
-};
 
 export default remarkCodeBlock;
